ContractClaimSystem.Tests/ClaimsControllerTest.cs
using Xunit;
using Moq;
using ContractClaimSystem.Controllers;
using ContractClaimSystem.Models;
using Microsoft.Extensions.Logging;
using Microsoft.AspNetCore.Mvc;
using Microsoft.EntityFrameworkCore;
using System.Threading.Tasks;
using System.Collections.Generic;
using Microsoft.AspNetCore.Identity;

public class ClaimsControllerTests
{
    private readonly ClaimsController _controller;
    private readonly Mock<ApplicationDbContext> _mockDbContext;
    private readonly Mock<UserManager<ApplicationUser>> _mockUserManager;
    private readonly Mock<ILogger<ClaimsController>> _mockLogger;

    public ClaimsControllerTests()
    {
        _mockDbContext = new Mock<ApplicationDbContext>();
        _mockUserManager = new Mock<UserManager<ApplicationUser>>(Mock.Of<IUserStore<ApplicationUser>>(), null, null, null, null, null, null, null, null);
        _mockLogger = new Mock<ILogger<ClaimsController>>();

        _controller = new ClaimsController(_mockDbContext.Object, _mockUserManager.Object, _mockLogger.Object);
    }

    [Fact]
    public async Task SubmitClaim_ReturnsRedirectToClaimSubmitted_OnValidClaim()
    {
        // Arrange
        var claim = new ClaimSubmission
        {
            HoursWorked = 10,
            HourlyRate = 20,
            AdditionalNotes = "Some notes"
        };

        // Act
        var result = await _controller.SubmitClaim(claim, null) as RedirectToActionResult;

        // Assert
        Assert.NotNull(result);
        Assert.Equal("ClaimSubmitted", result.ActionName);
    }

    [Fact]
    public async Task SubmitClaim_ReturnsView_OnInvalidModel()
    {
        // Arrange
        _controller.ModelState.AddModelError("Error", "Model Error");

        var claim = new ClaimSubmission
        {
            HoursWorked = 10
        };

        // Act
        var result = await _controller.SubmitClaim(claim, null) as ViewResult;

        // Assert
        Assert.NotNull(result);
        Assert.Equal(claim, result.Model);
    }

    // Additional tests for ManageClaims and others
}
